changed chat_id to bigint

+++ db/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, JSON, ForeignKey
+from sqlalchemy import Column, Integer, String, JSON, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 from db.database import Base
 from .mixins import Timestamp
@@ -8,7 +8,7 @@ class User(Timestamp, Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    chat_id = Column(Integer, unique=True)
+    chat_id = Column(BigInteger, unique=True)
     user_info = Column(JSON)
 
     message_history = relationship("MessageHistory", back_populates="owner")
